+++ app/src/main/kotlin/com/appspot/pistatium/dotsubscribe/MainActivity.kt
@@ -6,19 +6,16 @@ import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
-import android.util.Log
 import com.appspot.pistatium.dotsubscribe.adapters.ArticleAdapter
 import com.appspot.pistatium.dotsubscribe.databinding.ActivityMainBinding
 import com.appspot.pistatium.dotsubscribe.models.Article
+import com.appspot.pistatium.dotsubscribe.models.ArticleModel
 import com.appspot.pistatium.dotsubscribe.viewmodels.ContentViewModel
-import com.google.firebase.database.*
 import java.util.*
-import kotlin.concurrent.thread
 import kotlin.properties.Delegates
 
 
 class MainActivity : AppCompatActivity() {
-
 
     private var binding: ActivityMainBinding by Delegates.notNull()
 
@@ -29,22 +26,9 @@ class MainActivity : AppCompatActivity() {
         binding = DataBindingUtil.setContentView<ActivityMainBinding>(this, R.layout.activity_main)
         binding.contentViewModel = content
 
-        val database = FirebaseDatabase.getInstance()
-        // TODO:
-        val today = "20161210"
-        val current = "173000"
-        val latest_news = database.reference.child(today).child(current)
-        latest_news.addValueEventListener(object : ValueEventListener {
-            override fun onDataChange(snapshot: DataSnapshot) {
-                val articles =  ArrayList<Article>()
-                snapshot.children
-                        .map { it.getValue(Article::class.java) }
-                        .forEach { articles.add(it) }
-                content.articles = articles
-            }
-            override fun onCancelled(p0: DatabaseError?) { }
-
-        })
+        ArticleModel().getLatest().subscribe {
+            content.articles = ArrayList(it)
+        }
     }
 }
