+++ app/src/main/kotlin/com/appspot/pistatium/dotsubscribe/viewmodels/ContentViewModel.kt
@@ -8,7 +8,6 @@ import com.appspot.pistatium.dotsubscribe.models.Article
 import java.util.*
 
 
-
 class ContentViewModel(val context: Context) : BaseObservable() {
     @Bindable
     var articles = ArrayList<Article>()
